requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
joblib==1.3.2
openpyxl==3.1.2
scipy==1.11.4
